Refactor: all ids excl userId are uint64

+++ internal/model/time_record.go
@@ -6,9 +6,9 @@ import (
 )
 
 type TimeRecord struct {
-	ID        uint       `gorm:"primaryKey" json:"id"`
+	ID        uint64     `gorm:"primaryKey" json:"id"`
 	UserID    uuid.UUID  `gorm:"type:uuid;not null" json:"user_id"`
-	TaskID    int64      `gorm:"not null;index" json:"task_id"`
+	TaskID    uint64     `gorm:"not null;index" json:"task_id"`
 	StartTime time.Time  `gorm:"not null" json:"start_time"`
 	EndTime   *time.Time `json:"end_time,omitempty"`
 	IsClosed  bool       `gorm:"default:false" json:"is_closed"`
